Document how to add ShelveStorage in the code
The documentation for this is called 'writing a state storage'. 
I ignored this as I thought this was an advanced topic about extending Lamson.
Maybe there should be an alternative prototype 'config/prod.py' to hold a typical production setup, 
but for now I think if you grep for 'Shelve' or 'Storage' in your project, something should come back!